requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv~=1.0.1
pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0
future~=0.18.3
requests~=2.31.0
telebot~=0.0.5
peewee~=3.15.4
aiohttp
requests_async
httpx
peewee_async
tortoise-orm
aiosqlite